refactor: move email-smtp plugin to Node app
