+++ imports/node-app/plugins/email-smtp/index.js
@@ -1,7 +1,7 @@
-import mutations from "./mutations";
-import resolvers from "./resolvers";
-import schemas from "./schemas";
-import startup from "./startup";
+import mutations from "./mutations.js";
+import resolvers from "./resolvers.js";
+import schemas from "./schemas.js";
+import startup from "./startup.js";
 
 /**
  * @summary Import and call this function to add this plugin to your API.
@@ -12,7 +12,6 @@ export default async function register(app) {
   await app.registerPlugin({
     label: "SMTP Email",
     name: "reaction-email-smtp",
-    icon: "fa fa-envelope-o",
     functionsByType: {
       startup: [startup]
     },
