+++ imports/node-app/plugins/email-smtp/mutations/index.js
@@ -1,0 +1,5 @@
+import verifySMTPEmailSettings from "./verifySMTPEmailSettings.js";
+
+export default {
+  verifySMTPEmailSettings
+};
