+++ imports/node-app/plugins/email-smtp/resolvers/Mutation/index.js
@@ -1,0 +1,5 @@
+import verifySMTPEmailSettings from "./verifySMTPEmailSettings.js";
+
+export default {
+  verifySMTPEmailSettings
+};
