+++ imports/node-app/plugins/email-smtp/resolvers/index.js
@@ -1,0 +1,5 @@
+import Mutation from "./Mutation.js";
+
+export default {
+  Mutation
+};
