--- imports/plugins/included/email-smtp/server/no-meteor/mutations/index.js
@@ -1,5 +1,0 @@
-import verifySMTPEmailSettings from "./verifySMTPEmailSettings";
-
-export default {
-  verifySMTPEmailSettings
-};
