--- imports/plugins/included/email-smtp/server/no-meteor/resolvers/index.js
@@ -1,5 +1,0 @@
-import Mutation from "./Mutation";
-
-export default {
-  Mutation
-};
